Fix:MainInputPage수정#2
