+++ knock/src/pages/MainPage.tsx
@@ -6,35 +6,14 @@ import backgroundImg from '../core/assets/image/mainimage.png';
 import rightArrow from '../core/assets/icon/Hyphen-Arrow-right.svg';
 import styles from './MainPage.module.scss';
 import UButton from '../core/ui/buttons/UButton/UButton';
-import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { createSubject } from '../lib/api/Subject';
-import useUserInfo from '../lib/hooks/useUserInfo';
+import Icon from '../core/ui/CommonIcon/icon';
 
 const MainPage = () => {
-  const [name, setName] = useState('');
   const navigate = useNavigate();
-  const { handleUserInfo } = useUserInfo();
 
-  const handleNameSet = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setName(e.target.value);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (name && e.key === 'Enter') {
-      handleSubmit(e);
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLElement>) => {
-    e.preventDefault();
-    try {
-      const { id, name: subName, imageSource } = await createSubject({ name });
-      navigate(`/post/${id}/answer`);
-      handleUserInfo({ id, name: subName, imageSource });
-    } catch (e) {
-      console.error(e);
-    }
+  const goToAnswerPage = () => {
+    navigate(`/list`);
   };
 
   return (
@@ -42,15 +21,15 @@ const MainPage = () => {
       <MetaTags />
       <UButton
         type="box"
-        handleClick={handleSubmit}
+        handleClick={goToAnswerPage}
         isLightTheme={true}
         className={`${styles['mainPage__answer']}`}
       >
         답변하러 가기
-        <Image
+        <Icon
           src={rightArrow}
           alt="rightArrow"
-          imageClassName={styles['mainPage__answer__rightArrowImg']}
+          className={styles['mainPage__answer__rightArrowImg']}
         />
       </UButton>
       <Image
@@ -59,12 +38,7 @@ const MainPage = () => {
         imageClassName={styles['mainPage__logo']}
       />
       <div className={`${styles['mainPage__mainInput']}`}>
-        <MainPageInput
-          name={name}
-          handleNameSet={handleNameSet}
-          handleSubmit={handleSubmit}
-          handleKeyDown={handleKeyDown}
-        />
+        <MainPageInput />
       </div>
       <Image
         src={backgroundImg}
